Add files via upload
This is a MATLAB code for a Rock, Paper, Scissors, Lizard, Spock game app. It includes properties that correspond to different UI components, such as buttons and labels, as well as properties that store game data, such as the number of wins for the human and computer players. it was my first project in Matlab App Designer. 
note i did not include images in this repository